Align autocapture with Android: event data and dead click snapshot parity
- Coordinates: Double → Int in toProperties() to match JS SDK and Android
- Hash fallback ID: decimal → hex for shorter IDs matching Android format
- Dead click snapshot: add position/size to hash, switch from XOR to
  multiplicative (31*hash) for better collision resistance
- Consolidate duplicated SwiftUI detection and interaction checks into
  shared utilities in AutocaptureDefaults

+++ MixpanelDemo/MixpanelDemoTests/AutocaptureTests.swift
@@ -286,8 +286,8 @@ import XCTest
 
       let props = event.toProperties()
 
-      XCTAssertEqual(props["$x"] as? Double, 100)
-      XCTAssertEqual(props["$y"] as? Double, 200)
+      XCTAssertEqual(props["$x"] as? Int, 100)
+      XCTAssertEqual(props["$y"] as? Int, 200)
       XCTAssertEqual(props["$el_id"] as? String, "test_button")
       XCTAssertEqual(props["$el_tag_name"] as? String, "UIButton")
       XCTAssertEqual(props["$attr-aria-label"] as? String, "Test Button")
